Expose new repositoryType field for helm repositories
This is called "type" in flux -
https://fluxcd.io/docs/components/source/helmrepositories/#type - but
our frontend refers to kind as type, so I'm not taking any chances.

Flux always assumes "default" when it's not "oci", according to
https://github.com/fluxcd/source-controller/blob/812f6e49ddfbe42d5e11da2805b8feb9121cdb88/api/v1beta2/helmrepository_types.go#L34
This does the same thing - it means that if you run an older flux
that doesn't set this field, we will still set it, and set it to a
reasonable value.