documentation: Add titles and highlights to bits of getting started
Syntax highlighting is subtle, but pleasant.

When we say "make file, it should look like this", then putting the
filename in the title helps explaining which file we're talking about.

Using the `//highlight-start` and `//highlight-end` magic tags, it's
easier to say "this is the new bit, the rest is the old bit" which I'm
hoping makes it easier to see how to patch the podinfo kustomization.