documentation: Small word changes for clarity
The text explaining what the application view shows is very dense as
it is. Putting an example of a source kind in there was a good idea
when the source was displayed with its kind. Now that the prefix isn't
there, the example is one sub-clause too many.

The explanation of what sources are made my head spin. Try to break it
apart.

When we've just described all the tabs, there is no
`GitRepository/flux-system` link, so you can't click it.