Make the kinds allowed by getObject configurable
We should be able to see getObject for "any" object kind, so we can
display yaml for every object kind - however, we want to maintain some
kind of control to make sure that we don't serve up e.g. secrets.

Until we have the kind of control over this that we would want, this
dodgy setting lets every downstream user of this server just add them
to the list of "legal" object kinds, and then they'll show up.

This is supposed to be a relatively short-term fix to give the API
some amount of flexibility.