Remove final constant from old v1alpha1 CRD
This module used to contain the Application CRD. At the time the CRD
was removed, we left a few constants around that were used in many places.

Since then, there was only one constant left, and it was only used in
one place, so this moves that one to where it's used, so we can
finally finish the module removal.