Set version flags when building release images, too
When hitting `/v1/version` against a 0.9.0 image, it returns
`{"semver":"v0.0.0","commit":"","branch":"","buildTime":""}`. This is
incorrect.

This copy-pastes bits from the dev image build process that we use to
set these properly for the staging cluster.

The footer still prints the correct thing since the javascript does
have the right versions.