Use access token for UserInfo endpoint access
Some OIDC providers (such as Keycloak) check the token type to match
"Bearer" which is not true for the ID token which holds the token type
"ID". Therefore, the ID token cannot (and supposedly should not) be
used to authorize requests to resource servers.

With this change the auth server stores the access token alongside the
ID token in the browser's cookies so that it can be used by the
weave-gitops server later on, e.g. to make requests against the
UserInfo endpoint of the OIDC provider.

A caveat of this change is that the mock OIDC server used in tests
doesn't check the token type so without further changes to the
upstream library we cannot really test usage of the access token
properly.

closes #2125