Make `MultiAuthPrincipal` fail if none work
MultiAuthPrincipal is only used in the `WithAPIAuth` middleware and,
despite its name, doesn't just return a principal. How it is used means
that as well as returning a principal it implicitly is authenticating
that principal.

If no principal is found this should return an error. Not only has no
valid principal been found but, by implication of how this is used, the
provided authentication method has failed.