Add gzip compression
It was pointed out to me that weaveworks/weave-gitops-enterprise#989
had done this, and the results seemed promising (as, granted, adding
gzip usually does).

That PR only does it for static resources. I added it for API
endpoints too - the middleware is smart enough to check the response
size and only compress if it's large enough to make sense.