Fix the changelog generation
The changelog generation doesn't use the local git repository to find
the final tag, it uses the github API, so setting a tag in the local
repo doesn't help. Explicitly specify the toTag to be the current
commit instead.

However, doing that breaks the ignorePrereleases option so that the
release changelog only contains the changes since the last RC. Fix
that by adding a regex to filter out prereleases.