Remove the label error flow
A PR that's neither a feature or a bug, but that changes
e.g. dependencies, should still be in the changelog. The default
is "present, but not categorized", which is, like, actually a good
default.

This immediately makes me a lot less annoyed with the changelog,
causing me to think maybe we can just close #2279.