Stop creating resources in the hard-coded flux-system namespace
There's been a few weeks of backwards compatibility. Let's remove
that - if you're upgrading the chart, upgrade the image.

This makes no difference to any user who installs weave-gitops into
flux-system. This makes a difference to anybody who installs it in a
different namespace - the new chart will not be compatible with images
0.9.1 or before, however it will install half the number of roles,
bindings and secrets for anybody running a newer version than that.

This closes #2415.