gitops run: Don't abort uploading on error
Before this, if any upload threw any error, we stopped uploading - we
didn't shut anything down, we just stopped. When using this repository
as an example, running gitops run just prints:

```
► Refreshing bucket dev-bucket ...
Handling connection for 9000
....Handling connection for 9000
Handling connection for 9000
Handling connection for 9000
....
► Uploaded 88 files
✗ Error syncing directory: MissingContentLength
✗ Error syncing dir: MissingContentLength
► Request reconciliation of dev-bucket, and dev-ks (timeout 30s) ...
✗ Error requesting reconciliation: inventory is nil
✔ Reconciliation is done.
```
which is very confusing - I have about 100,000 files here.

This adds the filename to any failed upload - that way I can debug
what failed to upload.

This also completely swallows `MissingContentLength` errors - they
mean the upload had a length of 0, which is common (git doesn't
support checking in directories, so if you want to keep a directory
around you have to create an empty file and check that in).

And finally, when an upload fails, this continues - so bad
permissions, device files, massive images, or anything else that
causes a problem is reported, but doesn't interrupt the flow.