Enable stylecheck lint rule
This lint rule has some pretty good checks - for example the "error
strings shouldn't be capitalized", so we can always log errors
properly, or "duplicate imports" that make the code look more complex
than it is.

However, to enable it required me to search replace all sorts of
things to capitalize them: Url with URL, Uri with URI, Http with HTTP,
Json with JSON, etc.