Set gitops run chart version to "latest" when building CLI
We want the binary to install the version of the chart that matches
the version of the CLI at all times, so in order to achieve that this
we grab the most recent chart version from the chart when we build the
binary.

If you don't have yq installed locally, this will just be an empty
value, which means it'll install "latest". This is probably what you
want for development.

During the release process, this will first bump the chart version (in
prepare-release), then build the binary and release it using that
bumped chart version, and then release the chart. This is clumsy but
should mainly be the right order.