Use the correct labels to get reconciled objects
We didn't filter by namespace at all. This meant that we display all
automations in all namespaces that have been created by an automation
that matches this automation's name - e.g., the details page for the
helm release `weave-gitops` in namespace `prod` will show the
resources created by the helm release `weave-gitops` in namespace
`staging`. That's incorrect.

This was introduced in e8c89fb - before that we did filter by
namespace. That commit claims that it's done to make sure
`targetNamespace` works properly - I've created both kustomizations
and helm releases with targetNamespace set, and both of them work fine.

The reason is that e8c89fb contained 2 different changes - it also
stopped restricting the query to the namespace the automation was
installed into, which is what actually fixes the bug with targetNamespace.

This fixes #2761.