List automations objects (#2765)
* Make automation details use getObject instead of dedicated endpoints

This requires adding dedicated inventory handling for helm releases -
rather than redesigning the inventory handling, I added an inventory
field to the object envelope and only set it when the kind is
helmrelease. This isn't good, but also it works and maintains the old
behaviour.

There is one change in behaviour, which is that inventory read
failures from the inventory secret no longer causes the object to
disappear from lists or refuse to render details. It'll now show up,
you just won't see any of the objects it's created. This is required
for #2664 but as listObjects isn't used to list helmReleases yet, it
doesn't actually resolve it.

* useListAutomations

* first attempt to remove flux object kind

* change FluxObjectKind to Kind

* some lint and rebase errors

* more lint stuff

* literally just double quotes

* why is a different test failing every time HELP

* remove Kind specificity on requests

* remove Kind from Alert yaml view

* make kind string values consistent across backend using flux controller types

* missed a couple conversions in suspend_test.go

* accidental double import

Co-authored-by: Robin Sonefors <robin.sonefors@weave.works>