ui: Don't flatten children for the graph
I added the flattenChildren call after fetching reconciled objects in
da33a4b, where I assumed that ReconciliationGraph used getChildren
directly. Which it doesn't.

flattenChildren thus did it's thing of copying the children out to the
top layer so they show up in the Details view. And when the result of
that was displayed in the graph, each of these children would show up
both in the correct place, and at the top layer, and possibly in more
in-between layers.

This fixes it by introducing two different ways to get reconciled
objects - one that is flat and one that isn't.

I'm hoping we'll change from the tree structure to the d3-dag
structure one of these days, at which point it won't need to be nested
and this code can be removed.

This fixes #2808