Remove all context.Background() from pkg/run
Before this, hitting ctrl+c in the middle of syncing would just keep
printing error messages from the sync process about the bucket not
responding which is not exactly unexpected when the frontend process
has deleted the bucket.

Now hitting ctrl+c in the "wrong" place will interrupt uploading and
syncing - granted, it will still look ugly in places with messages
about "context cancelled", but it will stop immediately.