Add silly little struct for paths
I'm not sure how much of a good idea this actually is - it's
definitely under-developed. However, I had a small breakdown while
trying to get paths right and had to try something to sort some paths
out.

Basically, we expect the git repo to be "somewhere". Inside that git
repository, there's a manifests directory, which is the same as the
current working directory. Inside that directory, there's a clusters
directory. Somewhere, probably but not necessarily inside the manifests
directory but definitely inside the git directory, there's a workload
directory. All of these paths can be refered to as relative paths,
absolute paths, and some of them (in flux) should be prefixed with `./`.

This bans all somethingPath variables, and puts a paths variable
instead. That variable is a struct that has documented members so we
can remember which path is which. There's no variable called
`relativePath`.

Inside this struct, we should check those invariants described above,
but for the moment I mainly just copy-pasted the code we already had.