Put logr between cli logger and stdout
Before, we had a very simple log interface that wrote messages with
added emoji to a writer, in practice stdout.

This deletes the interface altogether - instead, the log functions are
now simply calls logr with an emoji prefix.

Then, it changes the fluxexec interface and the port forwarding
function to steal the stdout and stderr messages, and log them as logr
messages - info level for stdout, error level for stderr.

This means you can easily get out of the custom interface and switch
to using it as logr, with verbosity and structured output right there
- but you'd have to worry about any emoji yourself.

Still sent to stdout without going through a logger are the dots
during the upload progress (which, they're not log messages), and
interactive questions related to the dashboard or the bootstrap.

The one thing that doesn't work is when the CLI returns an error out
of cobra without logging it - we need to stop doing that.

To hook this up to something else, you need another zapcore Core
thingy that writes somewhere - to test this, I just wrote json to
stderr. I've left the code in there, commented out.

This fixes #2857.