Add support for listing objects filtered by labels
This is required to allow us to request gitops run sessions. It also
would let us make the detail view's labels clickable and take us to a
server-side filtered list view.

This requires listObject to use a POST request to be able to use a map
for labels in a compatible way - which looks a bit funny, but given
that all the clients are generated it doesn't matter that much.