use HTTPS in `gitops run` S3 bucket server
When installing the dev bucket server on the cluster, it is now
configured with a self-signed certificate generated on demand. The
certificate is valid for 3 days which should be enough for a typical
`gitops run` use case.

Instead of using HTTP in the forwarding from the host machine into the
bucket server pod, the CLI now uses HTTPS so that traffic can't be
intercepted on its way between the developer's machine and the
cluster.

The certificate, after being generated, is put into a Kubernetes
Secret which is in turn mounted into the bucket server pod.