fix npm timeouts in docker builds
We've been experiencing an increased number of build failures in the
release GitHub workflow, manifested by error messages like this:

```
> [linux/arm64 ui  9/11] RUN make node_modules:
 ERR! network Socket timeout
```

These failures only happen in the ARM64 builds which are executed in
a qemu emulation that apparently is so slow that it causes these
network timeouts. This is just a hypothesis, though, and this commit
is supposed to mitigate these timeouts by increasing npm's HTTP
timeouts.

See https://docs.npmjs.com/cli/v9/using-npm/config#fetch-timeout for
further documentation of the configuration parameter changed in this
commit.

closes #3229